+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@dataparty/dpc",
   "private": false,
-  "version": "0.4.0",
+  "version": "0.4.1",
   "description": "Dataparty cli",
   "author": "RosHub Inc. <code@roshub.io>",
   "license": "Apache-2.0",
