+++ src/utils/utils.js
@@ -1,0 +1,11 @@
+const uniqueArray = (arr)=>{
+  return arr.filter((v, i, a) => {
+    if( v !== undefined && a.indexOf(v) === i){
+      return true
+    }
+
+    return false
+  })
+}
+
+exports.uniqueArray = uniqueArray
